Move useEffect for signOut
