+++ src/pages/safe-list.tsx
@@ -1,7 +1,10 @@
 import MobileChatFooter from "@/components/chat/mobileChatFooter"
+import ViewCreateSafe from "@/components/chat/modals/CreateSafe"
 import { SafeList } from "@/components/chat/SafeList"
-import { getSession } from "next-auth/react"
-import React from "react"
+import useWallet from "@/hooks/wallets/useWallet"
+import { getSession, signOut } from "next-auth/react"
+import { useRouter } from "next/router"
+import React, { useEffect, useState } from "react"
 
 export async function getServerSideProps(context: any) {
 	const session = await getSession(context)
@@ -22,9 +25,21 @@ export async function getServerSideProps(context: any) {
 }
 
 const SafePage: React.FC<{ user: any }> = ({ user }) => {
+	const router = useRouter()
+	const wallet = useWallet()
+	const [createSafe, setCreateSafe] = useState<boolean>(false)
+	useEffect(() => {
+		if (user.address !== wallet?.address) {
+			signOut({ redirect: true })
+		}
+		if (router.asPath.includes('chain')) {
+			setCreateSafe(true)
+		}
+	}, [])
 	return (
 		<>
-			<SafeList user={user} />
+			{createSafe && <ViewCreateSafe open={createSafe} onClose={() => setCreateSafe(!createSafe)} />}
+			<SafeList createSafe={createSafe} setCreateSafe={setCreateSafe} />
 			<MobileChatFooter />
 		</>
 	)
